Add the iframe to the player map when it’s created
If the iframe was created by the library, it was storing the original
element and never the iframe. If you created another Player object after
the iframe was appended, it would select the iframe which wasn’t in the
map and so you’d get a new Player object that duplicated the message
listener.

Fixes #108